refactor: seperate node and utils file
