+++ ninshiki_yolo/utils/draw_detection_result.py
